Create ProtocolTestCase to contain protocol-only helpers
