+++ test/protocol/detach_raw_dap_test.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
 
-  class DetachTest1639218122 < TestCase
+  class DetachTest1639218122 < ProtocolTestCase
     PROGRAM = <<~RUBY
        1| module Foo
        2|   class Bar
