+++ test/protocol/detach_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
-  class DetachTest < TestCase
+  class DetachTest < ProtocolTestCase
     PROGRAM = <<~RUBY
        1| module Foo
        2|   class Bar
