+++ test/protocol/eval_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
-  class EvalTest < TestCase
+  class EvalTest < ProtocolTestCase
     PROGRAM = <<~RUBY
       1| a = 2
       2| b = 3
