+++ test/protocol/restart_raw_dap_test.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
 
-  class RestartTest1644070899 < TestCase
+  class RestartTest1644070899 < ProtocolTestCase
     PROGRAM = <<~RUBY
        1| module Foo
        2|    class Bar
