+++ test/protocol/step_back_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
-  class StepBackTest < TestCase
+  class StepBackTest < ProtocolTestCase
     PROGRAM = <<~RUBY
        1| binding.b do: 'record on'
        2|
