+++ test/protocol/step_raw_cdp_test.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
   
-  class StepTest1647142777 < TestCase
+  class StepTest1647142777 < ProtocolTestCase
     PROGRAM = <<~RUBY
       1| module Foo
       2|   class Bar
