+++ test/protocol/threads_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
-  class ThreadsTest < TestCase
+  class ThreadsTest < ProtocolTestCase
     PROGRAM = <<~RUBY
        1| def foo
        2|   Thread.new { sleep 30 }
