+++ test/protocol/watch_raw_dap_test.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require_relative '../support/test_case'
+require_relative '../support/protocol_test_case'
 
 module DEBUGGER__
 
-  class WatchTest1643810224 < TestCase
+  class WatchTest1643810224 < ProtocolTestCase
     PROGRAM = <<~RUBY
       1| a = 2
       2| a += 1
